requirements.txt
biopython==1.83         
python-dotenv==1.0.1    
ray[data,tune]==2.47.1     
orjson==3.10.2          
tqdm==4.66.4            
torch>=2.0.0
torchvision
torchaudio
transformers==4.41.2
datasets==2.19.1
accelerate==0.29.3
scikit-learn==1.5.0
numpy<2
numexpr